llm/requirements.txt
pandas
numpy
pydantic>=2
openai
python-dotenv
requests
streamlit
